Refactor tests based on the module they're testing
